chore(widgets): Light code reorg
