+++ modules/widgets/src/screenshot-widget.tsx
@@ -6,7 +6,7 @@
 import type {WidgetPlacement} from '@deck.gl/core';
 import {render} from 'preact';
 import {WidgetImpl, WidgetImplProps} from './widget-impl';
-import {IconButton} from './components';
+import {IconButton} from './lib/components';
 
 /** Properties for the ScreenshotWidget */
 export type ScreenshotWidgetProps = WidgetImplProps & {
