+++ modules/widgets/src/zoom-widget.tsx
@@ -12,7 +12,7 @@ import {
 } from '@deck.gl/core';
 import type {Deck, Viewport, Widget, WidgetPlacement} from '@deck.gl/core';
 import {render} from 'preact';
-import {ButtonGroup, GroupedIconButton} from './components';
+import {ButtonGroup, GroupedIconButton} from './lib/components';
 
 export type ZoomWidgetProps = {
   id?: string;
